Revert "Pin `jupyter_client` to `5.1.0` temporarily"
This reverts commit 4b033b7e4324612a298fae3c54f989c6da5c390a.

As the `jupyter_nbextensions_configurator` package in `conda-forge` has
since switched to using `pip` to do the install and fixed its entry
points to not use `setuptools`, there should no longer be an issue using
the latest `jupyter_client`. So drop this pinning so that the most
recent copy of `jupyter_client` is used and everything else that relies
upon it.