Make `/notebooks` directory in `for`-loop
As making the `/notebooks` directory outside the `for`-loop is causing
some errors to get suppressed due to the `;`. Simply make the
`/notebooks` directory in the `for`-loop and use `&&` to chain later
commands to it. This way if installing packages with `conda` runs into
issues, the failure will propagate to the Docker build and thus
terminate the build as expected.