Install ipywidgets
Including the `ipywidgets` package (formerly IPython's widgets now for
the Jupyter Notebook) to provide some added functionality with Jupyter
Notebooks. This is also used in the workflow, but it is nice to have
here to lighten the CI setup on the workflow's end. Also this is of
general utility for the user anyways.

Side note: Seems the post-link/pre-unlink steps now work in combination
with an activated `conda` environment. So we don't need to do those
manually to get this to work (verified by locally testing this), which
is great.