Install jupyter_contrib_nbextensions
Also include jupyter_contrib_nbextensions, which generally make working
with the Jupyter Notebook way more fun. Includes all sorts of options
including ones the user wasn't thinking of. Verified that installing in
the activated conda environment is sufficient for the
post-link/pre-unlink scripts to work providing a fully functional
install of the jupyter_contrib_nbextensions for the user's enjoyment.
Should also lighten the requirements in the workflow, but is still a
nice thing to have generally.