Replace `python3` with `python`
As activation of the `conda` environment now works correctly, there is
no need to specify that `python3` should be used. Instead by simply
specifying `python`, the right `python` from the active `conda`
environment (which will be the one with Python 3) will be used. So just
use `python` and let `conda` find the correct one.