Patch Python 2/3 kernels to use wrapper scripts
Currently the Jupyter Notebook will open each Python kernel, but not
activate their respective environment. This is a problem if the user
tries to run anything as a shell command as it will not run in the
correct environment. Thus it will pickup the wrong `pip` or `conda`
environment. In the best case, this results in an unusual result like
showing the wrong `python` is used. In the worst case, it could result
in changing the wrong environment or submitting info from the wrong
environment in a bug report.

To fix this, we install the `ipykernel` just as before, but we use a
hack to replace the `python` path to not point at the `python` in the
`conda` environment, but instead at the versioned `python` wrapper
script in `/usr/local/bin`. As that wrapper script activates the correct
`conda` environment before calling `python` to start the Jupyter
Notebook kernel, the user is place in a correctly activated environment.
All calls to `conda`, `python`, and `pip` will work as expected and
point at the correct executables. Thus along the Jupyter Notebook
kernels to seamless interoperate with the `conda` environment expected.
Should make debugging and further modification painless for the user.