Install the nbserverproxy conda-forge package too
This allows other neighboring port connections to the host machine
running the Jupyter notebook to be proxied as well. Can be helpful when
using the Dask Distributed's Bokeh-based Dashboard remotely (thus
avoiding an extra SSH port forwarding connection). Normally
`nbserverproxy` needs to be enabled as a Jupyter extension, but this
isn't needed with the `conda` package as this will be run as part of the
post-link script.