Pin Tornado to 5.0.2
Currently Tornado 5.1.0 breaks nbserverproxy. While there appears to be
a fix upstream, it has not yet been released. For now pin Tornado to the
latest of the 5.0.x series (i.e. 5.0.2), so that nbserverproxy can work
correctly. This pin can be released once a new release of nbserverproxy
comes out.