Create an entrypoint script to configure Jupyter
Make sure that Jupyter places the runtime directory in the user's home
directory. This is what Jupyter does by default if `XDG_RUNTIME_DIR` or
`JUPYTER_RUNTIME_DIR` are not set. So the behavior remains identical in
the Docker container with this change. This works nicely on the cluster
as well as it effectively overrides the location that `XDG_RUNTIME_DIR`
would have used. Not to mention this allows this information to be
persisted outside the container should it be useful. Will avoid needing
to muck with the user's `.bash_profile` on the cluster too much.